Move biweight tests
